Changed the nomos ui agents:
1. Changed to use the new sql proposed by Bob.  Tested, may need
changed in the backend nomos agent, as it never get's scheduled for
some reason, not sure if it's because a previous query is still using
jq_type.  I see no jq_type of 'nomos' so my theory is that is why it's
not getting scheduled by the ui, as that query is returning 0 rows.

2. changed the name from fo*whatever to nomos.  So ui-agents try to
sechedule a back end agent called nomos.  The names of the ui agents
where also changed.

agent-nomos-once.php works, tested, still puts out some debug info due
to the way nomos returns strings at this point.