Fix for defect 682.  If the Uri was not hostname/repo/?mod=refresh&remod=default then the simpleui url was not getting set.  Added code to set it when it was something other than ?mod=refresh&remod=default.
The other items mentioned in this defect have also been fixed.  The
Makefile was changed to use a posix contruct that will work
everywhere. This fixes the install error.

The infinite loop was caused by the simpleUi not being installed in a
source tree sandbox. New targets were also added to the makefiles
(srcinstall and srcuninstall) so the simpleUi and copyright would
work/install into a source tree.